Move timescale meter generation logic to useTimeScaleMeterConfiguration hook
