+++ src/components/Timeline/Timeline.tsx
@@ -28,10 +28,9 @@ import addHours from 'date-fns/addHours';
 import differenceInDays from 'date-fns/differenceInDays';
 import differenceInHours from 'date-fns/differenceInHours';
 import formatDate from 'date-fns/format';
-import getDaysInMonth from 'date-fns/getDaysInMonth';
 import isAfter from 'date-fns/isAfter';
 import isBefore from 'date-fns/isBefore';
-import { omit, result, uniqueId } from 'lodash';
+import { omit, result } from 'lodash';
 import {
   Fragment,
   ReactElement,
@@ -59,15 +58,11 @@ import {
   useScrollTimelineTools,
   useTimeScaleTool,
 } from './hooks';
+import { useTimeScaleMeterConfiguration } from './hooks/TimeScaleMeterConfiguration';
 import {
   ScrollToDateFunction,
   ScrollToDateFunctionOptions,
-  TimeScaleConfiguration,
   TimeScaleOption,
-  TimeScaleRow,
-  fullMonthLabels,
-  quarterLabels,
-  shortMonthLabels,
   timeScaleOptions,
   timelineSearchParamValidationSpec,
 } from './models';
@@ -878,6 +873,16 @@ export const BaseTimeline = <RecordRow extends BaseDataRow>(
   }, [scrollToDate]);
 
   const selectedTimeScale = ((): TimeScaleOption => {
+    if (
+      isCustomDatesSelected &&
+      customDateRange?.startDate &&
+      customDateRange?.endDate
+    ) {
+      return getIdealOptimalTimeScale({
+        startDate: createDateWithoutTimezoneOffset(customDateRange.startDate),
+        endDate: createDateWithoutTimezoneOffset(customDateRange.endDate),
+      });
+    }
     if (searchParamsSelectedTimeScale) {
       return searchParamsSelectedTimeScale;
     }
@@ -999,490 +1004,17 @@ export const BaseTimeline = <RecordRow extends BaseDataRow>(
     );
   }, [customDateRange, isCustomDatesSelected]);
 
-  const getTimeScaleMeterConfiguration = useCallback(
-    (selectedTimeScale: TimeScaleOption) => {
-      const {
-        timeScaleRows,
-        unitTimeScaleWidth: baseUnitTimeScaleWidth,
-        timeScaleWidth,
-      } = ((): TimeScaleConfiguration => {
-        const getDailyTickTimeScale = ({
-          dayWidth,
-          unitTimeScale,
-          dayOfWeekFormat,
-        }: {
-          dayWidth: number;
-          unitTimeScale: number;
-          dayOfWeekFormat: string;
-        }): TimeScaleConfiguration => {
-          return {
-            timeScaleRows: [
-              timelineYears.flatMap((year) => {
-                return fullMonthLabels.map((monthLabel) => {
-                  return {
-                    id: uniqueId(),
-                    label: `${monthLabel} ${year}`,
-                  };
-                });
-              }),
-              (() => {
-                if (TimeScaleMeterPropsVariant === 'compact') {
-                  return timelineYears.flatMap((year) => {
-                    return fullMonthLabels.flatMap((_, index) => {
-                      const unitTickDate = new Date(year, index, 1);
-                      const daysInMonth = getDaysInMonth(unitTickDate);
-                      return Array.from({ length: daysInMonth }).map(
-                        (_, dayOfMonthIndex) => {
-                          const tickDate = addDays(
-                            unitTickDate,
-                            dayOfMonthIndex
-                          );
-                          const dayOfWeekIndex = tickDate.getDay();
-                          return {
-                            id: uniqueId(),
-                            label: formatDate(
-                              tickDate,
-                              dayOfMonthIndex === 0
-                                ? 'MMM d, yyyy'
-                                : 'EEE, MMM d'
-                            ),
-                            showLabel:
-                              dayOfMonthIndex === 0 ||
-                              (dayOfWeekIndex === 1 &&
-                                dayOfMonthIndex >= 2 &&
-                                daysInMonth - dayOfMonthIndex >= 2),
-                            color: alpha(palette.text.primary, 0.3),
-                            ...(() => {
-                              if (dayOfMonthIndex === 0) {
-                                return {
-                                  color: palette.text.primary,
-                                };
-                              }
-                            })(),
-                          } as TimeScaleRow;
-                        }
-                      );
-                    });
-                  });
-                }
-                return timelineYears.flatMap((year) => {
-                  return fullMonthLabels.flatMap((_, index) => {
-                    const unitTickDate = new Date(year, index, 1);
-                    const daysInMonth = getDaysInMonth(unitTickDate);
-                    return Array.from({ length: daysInMonth }).map(
-                      (_, index) => {
-                        const tickDate = addDays(unitTickDate, index);
-                        return {
-                          id: uniqueId(),
-                          label: formatDate(tickDate, dayOfWeekFormat),
-                        };
-                      }
-                    );
-                  });
-                });
-              })(),
-              timelineYears.flatMap((year) => {
-                return fullMonthLabels.flatMap((_, index) => {
-                  const unitTickDate = new Date(year, index, 1);
-                  const daysInMonth = getDaysInMonth(unitTickDate);
-                  return Array.from({ length: daysInMonth }).map(
-                    (_, dayOfMonthIndex) => {
-                      const tickDate = addDays(unitTickDate, dayOfMonthIndex);
-                      const dayOfWeekIndex = tickDate.getDay();
-                      return {
-                        id: uniqueId(),
-                        label: formatDate(tickDate, 'd'),
-                        ...(() => {
-                          if (TimeScaleMeterPropsVariant === 'compact') {
-                            return {
-                              sx: {
-                                ...(() => {
-                                  if (dayOfWeekIndex === 1) {
-                                    return {
-                                      borderLeftColor: '#f00',
-                                    };
-                                  }
-                                })(),
-                                ...(() => {
-                                  if (dayOfMonthIndex === 0) {
-                                    return {
-                                      borderLeftColor: palette.text.primary,
-                                    };
-                                  }
-                                })(),
-                              },
-                            };
-                          }
-                        })(),
-                      };
-                    }
-                  );
-                });
-              }),
-            ],
-            unitTimeScaleWidth: dayWidth * unitTimeScale,
-            timeScaleWidth: totalNumberOfDays * dayWidth,
-          };
-        };
-
-        const getMonthlyTickTimeScale = ({
-          monthSplit,
-          unitTimeScaleWidth,
-          timeScaleWidth,
-        }: {
-          monthSplit: number;
-          unitTimeScaleWidth: number;
-          timeScaleWidth: number;
-        }): TimeScaleConfiguration => {
-          const totalTimeScaleRegions = timelineYears.length * 12 * monthSplit;
-          return {
-            timeScaleRows: [
-              timelineYears.flatMap((year) => {
-                return quarterLabels.map((quarter) => {
-                  const label = `${quarter} ${year}`;
-                  return {
-                    id: uniqueId(),
-                    label,
-                  };
-                });
-              }),
-              (() => {
-                if (TimeScaleMeterPropsVariant === 'compact') {
-                  return timelineYears.flatMap((year) => {
-                    return fullMonthLabels.map((label, monthIndex) => {
-                      return {
-                        id: uniqueId(),
-                        label: (() => {
-                          if (
-                            selectedTimeScale === 'Quarter' ||
-                            monthIndex === 0
-                          ) {
-                            if (
-                              selectedTimeScale === 'Quarter' &&
-                              monthIndex % 3 === 0
-                            ) {
-                              return `Q${
-                                Math.floor(monthIndex / 3) + 1
-                              } ${label.slice(0, 3)} ${year}`;
-                            }
-                            return `${label.slice(0, 3)} ${year}`;
-                          }
-                          return label.slice(0, 3);
-                        })(),
-                        sx: {
-                          color: alpha(palette.text.primary, 0.3),
-                          ...(() => {
-                            if (
-                              (selectedTimeScale === 'Quarter' &&
-                                monthIndex % 3 === 0) ||
-                              monthIndex === 0
-                            ) {
-                              return {
-                                color: palette.text.primary,
-                              };
-                            }
-                          })(),
-                        },
-                      } as TimeScaleRow;
-                    });
-                  });
-                }
-                return timelineYears.flatMap(() => {
-                  return fullMonthLabels.map((label) => {
-                    return {
-                      id: uniqueId(),
-                      label,
-                    };
-                  });
-                });
-              })(),
-              timelineYears.flatMap((_, yearIndex) => {
-                return Array.from({ length: 12 }).flatMap((_, monthIndex) => {
-                  return Array.from({ length: monthSplit }).map(
-                    (_, periodIndex) => {
-                      const unitTickDate = addDays(
-                        minCalendarDate,
-                        Math.round(
-                          totalNumberOfDays *
-                            (((yearIndex * 12 + monthIndex) * monthSplit +
-                              periodIndex) /
-                              totalTimeScaleRegions)
-                        )
-                      );
-                      return {
-                        id: uniqueId(),
-                        label: unitTickDate.getDate(),
-                        sx: {
-                          ...(() => {
-                            if (
-                              (selectedTimeScale === 'Quarter' &&
-                                monthIndex % 3 === 0 &&
-                                periodIndex === 0) ||
-                              (monthIndex === 0 && periodIndex === 0)
-                            ) {
-                              return {
-                                borderLeftColor: palette.text.primary,
-                              };
-                            }
-                          })(),
-                        },
-                      };
-                    }
-                  );
-                });
-              }),
-            ],
-            unitTimeScaleWidth,
-            timeScaleWidth,
-          };
-        };
-
-        switch (selectedTimeScale) {
-          case 'Day':
-            const hourWidth = 64;
-            return {
-              timeScaleRows: [
-                timelineYears.flatMap((year) => {
-                  return fullMonthLabels.flatMap((_, monthIndex) => {
-                    return Array.from({
-                      length: getDaysInMonth(new Date(year, monthIndex, 1)),
-                    }).map((_, dayIndex) => {
-                      return {
-                        id: uniqueId(),
-                        label: formatDate(
-                          new Date(year, monthIndex, dayIndex + 1),
-                          'MMMM d, yyyy'
-                        ),
-                      };
-                    });
-                  });
-                }),
-                timelineYears.flatMap((year) => {
-                  return fullMonthLabels.flatMap((_, monthIndex) => {
-                    return Array.from({
-                      length: getDaysInMonth(new Date(year, monthIndex, 1)),
-                    }).map((_, dayIndex) => {
-                      return {
-                        id: uniqueId(),
-                        label: formatDate(
-                          new Date(year, monthIndex, dayIndex + 1),
-                          'EEEE'
-                        ),
-                      };
-                    });
-                  });
-                }),
-                timelineYears.flatMap((year) => {
-                  return fullMonthLabels.flatMap((_, monthIndex) => {
-                    return Array.from({
-                      length: getDaysInMonth(new Date(year, monthIndex, 1)),
-                    }).flatMap((_, dayIndex) => {
-                      const tickDate = new Date(year, monthIndex, dayIndex + 1);
-                      return Array.from({ length: 24 }).map((_, hourIndex) => {
-                        return {
-                          id: uniqueId(),
-                          label: formatDate(
-                            addHours(tickDate, hourIndex),
-                            'h a'
-                          ),
-                        };
-                      });
-                    });
-                  });
-                }),
-              ],
-              unitTimeScaleWidth: 24 * hourWidth,
-              timeScaleWidth: totalNumberOfHours * hourWidth,
-            };
-          case 'Week':
-            return getDailyTickTimeScale({
-              dayOfWeekFormat: 'EEEE',
-              dayWidth: 200,
-              unitTimeScale: 7,
-            });
-          case '2 week':
-            return getDailyTickTimeScale({
-              dayOfWeekFormat: 'EEE',
-              dayWidth: 100,
-              unitTimeScale: 15,
-            });
-          case 'Month':
-            return getDailyTickTimeScale({
-              dayOfWeekFormat: 'EEEEE',
-              dayWidth: 60,
-              unitTimeScale: 30,
-            });
-          case 'Quarter':
-            return getMonthlyTickTimeScale({
-              monthSplit: 4,
-              unitTimeScaleWidth: 480 * 3,
-              timeScaleWidth: 480 * 12 * timelineYears.length,
-            });
-          case 'Year':
-            return getMonthlyTickTimeScale({
-              monthSplit: (() => {
-                if (TimeScaleMeterPropsVariant === 'compact') {
-                  return 1;
-                }
-                return 3;
-              })(),
-              unitTimeScaleWidth: 120 * 12,
-              timeScaleWidth: 120 * 12 * timelineYears.length,
-            });
-          case '5 year': {
-            const yearWidth = 360;
-            const unitTimeScaleWidth = yearWidth * 5;
-            return {
-              timeScaleRows: [
-                timelineYears.flatMap((year) => {
-                  return {
-                    id: uniqueId(),
-                    label: String(year),
-                  };
-                }),
-                (() => {
-                  if (TimeScaleMeterPropsVariant === 'compact') {
-                    return timelineYears.flatMap((year) => {
-                      return {
-                        id: uniqueId(),
-                        label: String(year),
-                      };
-                    });
-                  }
-                  return timelineYears.flatMap(() => {
-                    return quarterLabels.map((quarter) => {
-                      return {
-                        id: uniqueId(),
-                        label: quarter,
-                      };
-                    });
-                  });
-                })(),
-                timelineYears.flatMap(() => {
-                  return shortMonthLabels.map((label, monthIndex) => {
-                    return {
-                      id: uniqueId(),
-                      label,
-                      ...(() => {
-                        if (TimeScaleMeterPropsVariant === 'compact') {
-                          return {
-                            sx: {
-                              ...(() => {
-                                if (monthIndex % 3 === 0) {
-                                  return {
-                                    borderLeftColor: '#f00',
-                                  };
-                                }
-                              })(),
-                              ...(() => {
-                                if (monthIndex === 0) {
-                                  return {
-                                    borderLeftColor: palette.text.primary,
-                                  };
-                                }
-                              })(),
-                            },
-                          };
-                        }
-                      })(),
-                    };
-                  });
-                }),
-              ],
-              unitTimeScaleWidth,
-              timeScaleWidth: yearWidth * timelineYears.length,
-            };
-          }
-        }
-      })();
-
-      const unitTimeScaleWidth = (() => {
-        if (
-          isCustomDatesSelected &&
-          customDateRange?.startDate &&
-          customDateRange?.endDate
-        ) {
-          const timeScaleHourWidth = (() => {
-            switch (selectedTimeScale) {
-              case 'Day':
-                return baseUnitTimeScaleWidth / 24;
-              case 'Week':
-                return baseUnitTimeScaleWidth / (7 * 24);
-              case '2 week':
-                return baseUnitTimeScaleWidth / (2 * 7 * 24);
-              case 'Month':
-                return baseUnitTimeScaleWidth / (30 * 24);
-              case 'Quarter':
-                return baseUnitTimeScaleWidth / (3 * 30 * 24);
-              case 'Year':
-                return baseUnitTimeScaleWidth / (365 * 24);
-              case '5 year':
-                return baseUnitTimeScaleWidth / (5 * 365 * 24);
-            }
-          })();
-          return (
-            timeScaleHourWidth *
-            differenceInHours(
-              createDateWithoutTimezoneOffset(customDateRange.endDate),
-              createDateWithoutTimezoneOffset(customDateRange.startDate)
-            )
-          );
-        }
-        return baseUnitTimeScaleWidth;
-      })();
-
-      return {
-        timeScaleRows,
-        unitTimeScaleWidth,
-        timeScaleWidth,
-      };
-    },
-    [
-      TimeScaleMeterPropsVariant,
-      customDateRange?.endDate,
-      customDateRange?.startDate,
-      isCustomDatesSelected,
+  const { timeScaleRows, unitTimeScaleWidth, timeScaleWidth } =
+    useTimeScaleMeterConfiguration({
+      selectedTimeScale,
       minCalendarDate,
-      palette.text.primary,
       timelineYears,
+      TimeScaleMeterPropsVariant,
       totalNumberOfDays,
       totalNumberOfHours,
-    ]
-  );
-  const getTimeScaleMeterConfigurationRef = useRef(
-    getTimeScaleMeterConfiguration
-  );
-  getTimeScaleMeterConfigurationRef.current = getTimeScaleMeterConfiguration;
-
-  const {
-    timeScaleRows: baseTimeScaleRows,
-    unitTimeScaleWidth,
-    timeScaleWidth,
-  } = useMemo(() => {
-    return getTimeScaleMeterConfiguration(selectedTimeScale);
-  }, [getTimeScaleMeterConfiguration, selectedTimeScale]);
-
-  const timeScaleRows = useMemo(() => {
-    if (
-      isCustomDatesSelected &&
-      customDateRange?.startDate &&
-      customDateRange.endDate
-    ) {
-      const { timeScaleRows } = getTimeScaleMeterConfigurationRef.current(
-        getIdealOptimalTimeScaleRef.current({
-          startDate: createDateWithoutTimezoneOffset(customDateRange.startDate),
-          endDate: createDateWithoutTimezoneOffset(customDateRange.endDate),
-        })
-      );
-      return timeScaleRows;
-    }
-    return baseTimeScaleRows;
-  }, [
-    baseTimeScaleRows,
-    customDateRange?.endDate,
-    customDateRange?.startDate,
-    isCustomDatesSelected,
-  ]);
+      customDateRange,
+      isCustomDatesSelected,
+    });
 
   //#region Unit time scaling
   const [timelineWidthScaleFactor, setTimelineWidthScaleFactor] = useState(
@@ -2016,7 +1548,7 @@ export const BaseTimeline = <RecordRow extends BaseDataRow>(
 
   const dateAtCursorBgcolor = (palette.mode === 'light' ? lighten : darken)(
     palette.text.primary,
-    0.5
+    0.75
   );
   const columns: TableColumn<RecordRow>[] = [
     {
@@ -2122,55 +1654,55 @@ export const BaseTimeline = <RecordRow extends BaseDataRow>(
             }}
           >
             {(() => {
-              if (isCustomDatesSelected) {
+              if (
+                isCustomDatesSelected &&
+                customDateRange?.startDate &&
+                customDateRange.endDate
+              ) {
                 const bgcolor = alpha(palette.grey[900], 0.75);
                 const borderColor = alpha('#f00', 0.15);
                 return (
                   <>
-                    {customDateRange?.startDate ? (
-                      <Box
-                        className={clsx(classes.customDateRangeBlocker)}
-                        sx={{
-                          position: 'absolute',
-                          top: 0,
-                          left: -baseSpacingUnits,
-                          width: `calc(${
-                            getPercentageAtDate(
-                              createDateWithoutTimezoneOffset(
-                                customDateRange.startDate
-                              )
-                            ) * 100
-                          }% + ${baseSpacingUnits}px)`,
-                          bgcolor,
-                          borderColor,
-                          borderRight: `1px solid ${borderColor}`,
-                          pointerEvents: 'none',
-                          zIndex: 3,
-                        }}
-                      />
-                    ) : null}
-                    {customDateRange?.endDate ? (
-                      <Box
-                        className={clsx(classes.customDateRangeBlocker)}
-                        sx={{
-                          position: 'absolute',
-                          top: 0,
-                          right: -baseSpacingUnits,
-                          left: `${
-                            getPercentageAtDate(
-                              createDateWithoutTimezoneOffset(
-                                customDateRange.endDate
-                              )
-                            ) * 100
-                          }%`,
-                          bgcolor,
-                          borderColor,
-                          borderLeft: `1px solid ${borderColor}`,
-                          pointerEvents: 'none',
-                          zIndex: 3,
-                        }}
-                      />
-                    ) : null}
+                    <Box
+                      className={clsx(classes.customDateRangeBlocker)}
+                      sx={{
+                        position: 'absolute',
+                        top: 0,
+                        left: -baseSpacingUnits,
+                        width: `calc(${
+                          getPercentageAtDate(
+                            createDateWithoutTimezoneOffset(
+                              customDateRange.startDate
+                            )
+                          ) * 100
+                        }% + ${baseSpacingUnits}px)`,
+                        bgcolor,
+                        borderColor,
+                        borderRight: `1px solid ${borderColor}`,
+                        pointerEvents: 'none',
+                        zIndex: 3,
+                      }}
+                    />
+                    <Box
+                      className={clsx(classes.customDateRangeBlocker)}
+                      sx={{
+                        position: 'absolute',
+                        top: 0,
+                        right: -baseSpacingUnits,
+                        left: `${
+                          getPercentageAtDate(
+                            createDateWithoutTimezoneOffset(
+                              customDateRange.endDate
+                            )
+                          ) * 100
+                        }%`,
+                        bgcolor,
+                        borderColor,
+                        borderLeft: `1px solid ${borderColor}`,
+                        pointerEvents: 'none',
+                        zIndex: 3,
+                      }}
+                    />
                   </>
                 );
               }
